feat(css-injector): simplify implementation
