+++ packages/x-archetype-utils/src/index.ts
@@ -1,6 +1,4 @@
 export * from './build/rollup/rollup.config'
 export * from './build/webpack/webpack.config'
-export * from './css-injector/css-injector'
-export * from './css-injector/css-injector.types'
 export * from './i18n/i18n.plugin'
 export * from './i18n/i18n.types'
